requirements.txt
arabic-reshaper==3.0.0
colorama==0.4.6
numpy==1.21.6
opencv-python==4.9.0.80
Pillow==9.5.0
pypng==0.20220715.0
PyQRCode==1.2.1
python-bidi==0.4.2
pyzbar==0.1.9
qrcode==7.4.2
six==1.16.0
typing_extensions==4.7.1
